Move sample code file to default module
