+++ java-checks/src/test/java/org/sonar/java/checks/UtilityClassWithPublicConstructorCheckTest.java
@@ -22,12 +22,14 @@ package org.sonar.java.checks;
 import org.junit.jupiter.api.Test;
 import org.sonar.java.checks.verifier.CheckVerifier;
 
+import static org.sonar.java.checks.verifier.TestUtils.mainCodeSourcesPath;
+
 class UtilityClassWithPublicConstructorCheckTest {
 
   @Test
   void test() {
     CheckVerifier.newVerifier()
-      .onFile("src/test/files/checks/UtilityClassWithPublicConstructorCheck.java")
+      .onFile(mainCodeSourcesPath("checks/UtilityClassWithPublicConstructorCheckSample.java"))
       .withCheck(new UtilityClassWithPublicConstructorCheck())
       .verifyIssues();
   }
